requirements.txt
Django==2.2.10
dnspython==1.16.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
protobuf==3.6.1
PyMySQL==0.9.3
pytz==2019.1
six==1.14.0
sqlparse==0.3.0
virtualenv==16.5.0
coverage==5.1
coveralls==2.0.0
django-cors-headers==3.2.1
python-dateutil==2.8.1

